Add RegexExpr.literal convenience constructor
This alternative constructor serves two purposes: firstly it is provided as a
convenience but secondly it makes it easier for literal regexes to be
consistently created. For example, re.escape will pass "foo!" as-is in new
versions of Python but replace it with r"foo\!" in older ones. Using this
constructor, differences in behaviour can be glossed over. This change enables
the meta_grammar equivalence test to pass correctly under both Python 3.6 and
3.7+.